pyproject.toml
[project]
name = "scout"
version = "0.1.0"
description = "Self-hosted data agent platform for AI-powered database querying"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Django
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-environ>=0.11",
    "psycopg2-binary>=2.9",
    # Auth
    "django-allauth>=65.0",
    "PyJWT[crypto]>=2.0",
    "cryptography>=42.0",
    # LangGraph / LangChain
    "langgraph>=0.2",
    "langchain-anthropic>=0.2",
    "langchain-core>=0.3",
    "langgraph-checkpoint-postgres>=2.0",
    # SQL validation
    "sqlglot>=25.0",
    # Visualization & data
    "plotly>=5.0",
    "pandas>=2.0",
    "kaleido>=0.2",
    "sqlalchemy>=2.0",
    # Security / HTML sanitization
    "bleach>=6.0",
    # Caching & Task Queue
    "redis>=5.0",
    "celery[redis]>=5.4",
    "django-celery-beat>=2.6",
    # Async database (required by langgraph-checkpoint-postgres)
    "asyncpg>=0.29",
    # Utilities
    "python-dotenv>=1.0",
    # Production server
    "gunicorn>=21.0",
    "uvicorn[standard]>=0.30",
    "pyyaml>=6.0.3",
    # MCP server
    "mcp>=1.0",
    "langchain-mcp-adapters>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "factory-boy>=3.3",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config", "mcp_server"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"config/settings/*.py" = ["E402", "F401", "F403", "F405"]
"scripts/*.py" = ["E402"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
    "zensical>=0.0.21",
]
